displaying api data
Edited home_view.dart
Edited product_model.dart
Viewed custom_card.dart:24-42
Edited custom_card.dart
Edited custom_card.dart
Ran command: `git status`
Ran command: `git diff`

```git
fix: bind product data to UI and fix rendering issues

- Connect `FutureBuilder` in `HomeView` to `GetAllProductsService` with error handling
- Update `CustomCard` to accept and display dynamic `ProductModel` data
- Add text overflow truncation and image dimension constraints in `CustomCard`
- Ensure `price` double conversion in `ProductModel.fromJson`
```

+++ lib/models/product_model.dart
@@ -23,7 +23,7 @@ class ProductModel {
     return ProductModel(
       id: jsonData['id'],
       title: jsonData['title'],
-      price: jsonData['price'],
+      price: jsonData['price'].toDouble(),
       description: jsonData['description'],
       category: jsonData['category'],
       image: jsonData['image'],
